chore: update template to use Remix Vite
Remix sites can either use Remix Vite (now the default) or the Remix
"Classic Compiler".

Hydrogen sites have supported Remix Vite since ~April 2024 and this is
now the default for Hydrogen sites.

This updates the Netlify template to use Remix Vite.

It also updates it to the latest version of Hydrogen and pulls in
upstream updates to the core Hydrogen template.