feat: add full page caching
Cache most pages for 1 hour on the Netlify CDN with stale-while-revalidate (allowing 1 min for
background revalidation to complete). Cache these also on browsers (and anything else upstream) but
require a conditional request to revalidate.

On the search page, vary on the search input.

Disable this caching on pages that have personalization: account/*, cart.